Struct transit_grid::graphs::TransitNetwork

pub struct TransitNetwork<R: Copy, T: CoordNum> {
    pub physical_graph: PhysicalGraph<R, T>,
    pub topology_graph: TopologyGraph,
}

Represents a transit network as a graph with transit nodes and edges.

The struct holds a physical graph and a topological graph which are lower-level representations of the network. TransitNetwork provides a higher-level interface to the physical graph and topological graph.

The struct implements TransitNetworkModifier trait for modifying the underlying physical graph.

Generics

R: Copyable trait bound. This represents the type of the data associated with the network’s routes. T: This represents the type of the coordinates used in the network. It’s expected to implement CoordNum trait.

Fields

physical_graph: PhysicalGraph<R, T> - The physical graph representing the transit network.

topology_graph: TopologyGraph - The topological graph representing the transit network.

Fields
physical_graph: PhysicalGraph<R, T>

The physical graph representing the transit network.

topology_graph: TopologyGraph

The topological graph representing the transit network.

Implementations

impl<R: Copy, T: CoordNum> TransitNetwork<R, T>

pub fn new() -> Self

Constructs a new TransitNetwork with an empty PhysicalGraph and TopologyGraph.

Returns

A new TransitNetwork instance.

pub fn get_edge_by_id(&self, edge_id: EdgeId) -> Option<&TransitEdge<T>>

Returns a reference to the TransitNode with the given ID.